+++ cmd/internal/cli/build.go
@@ -55,8 +55,10 @@ var buildArgs struct {
 	userns              bool     // Enable user namespaces
 	ignoreSubuid        bool     // Ignore /etc/subuid entries (hidden)
 	ignoreFakerootCmd   bool     // Ignore fakeroot command (hidden)
+	ignoreProot         bool     // Ignore proot command (hidden)
 	ignoreUserns        bool     // Ignore user namespace(hidden)
 	remote              bool     // Remote flag(hidden, only for helpful error message)
+	reproducible        bool     // Reproducible build
 	buildVarArgs        []string // Variables passed to build procedure.
 	buildVarArgFile     string   // Variables file passed to build procedure.
 	buildArgsUnusedWarn bool     // Variables passed to build procedure to turn fatal error to warn.
@@ -313,6 +315,17 @@ var buildIgnoreFakerootCommand = cmdline.Flag{
 	Hidden:       true,
 }
 
+// --ignore-proot
+var buildIgnoreProot = cmdline.Flag{
+	ID:           "buildIgnoreProotFlag",
+	Value:        &buildArgs.ignoreProot,
+	DefaultValue: false,
+	Name:         "ignore-proot",
+	Usage:        "ignore proot",
+	EnvKeys:      []string{"IGNORE_PROOT"},
+	Hidden:       true,
+}
+
 // --ignore-userns
 var buildIgnoreUsernsFlag = cmdline.Flag{
 	ID:           "buildIgnoreUsernsFlag",
@@ -332,6 +345,15 @@ var buildRemoteFlag = cmdline.Flag{
 	Usage:        "--remote is no longer supported, try building locally without it",
 	EnvKeys:      []string{},
 	Hidden:       true,
+}
+
+var buildReproducibleFlag = cmdline.Flag{
+	ID:           "buildReproducibleFlag",
+	Value:        &buildArgs.reproducible,
+	DefaultValue: false,
+	Name:         "reproducible",
+	Usage:        "creates a reproducible build by using the creation date of the source image",
+	EnvKeys:      []string{"REPRODUCIBLE"},
 }
 
 // --build-arg
@@ -402,8 +424,10 @@ func init() {
 		cmdManager.RegisterFlagForCmd(&buildUsernsFlag, buildCmd)
 		cmdManager.RegisterFlagForCmd(&buildIgnoreSubuidFlag, buildCmd)
 		cmdManager.RegisterFlagForCmd(&buildIgnoreFakerootCommand, buildCmd)
+		cmdManager.RegisterFlagForCmd(&buildIgnoreProot, buildCmd)
 		cmdManager.RegisterFlagForCmd(&buildIgnoreUsernsFlag, buildCmd)
 		cmdManager.RegisterFlagForCmd(&buildRemoteFlag, buildCmd)
+		cmdManager.RegisterFlagForCmd(&buildReproducibleFlag, buildCmd)
 
 		cmdManager.RegisterFlagForCmd(&buildVarArgsFlag, buildCmd)
 		cmdManager.RegisterFlagForCmd(&buildVarArgFileFlag, buildCmd)