+++ cmd/internal/cli/build_linux.go
@@ -203,6 +203,10 @@ func runBuild(cmd *cobra.Command, args []string) {
 		}
 	}
 
+	if buildArgs.ignoreProot {
+		os.Setenv("APPTAINER_IGNORE_PROOT", "1")
+	}
+
 	if buildArgs.nvidia {
 		os.Setenv("APPTAINER_NV", "1")
 	}
@@ -370,39 +374,39 @@ func runBuildLocal(ctx context.Context, cmd *cobra.Command, dst, spec string, fa
 		sylog.Fatalf("%v", err)
 	}
 
-	b, err := build.New(
-		defs,
-		build.Config{
-			Dest:      dst,
-			Format:    buildFormat,
-			NoCleanUp: buildArgs.noCleanUp,
-			Opts: types.Options{
-				ImgCache:          imgCache,
-				TmpDir:            tmpDir,
-				NoCache:           disableCache,
-				Update:            buildArgs.update,
-				Force:             forceOverwrite,
-				Sections:          buildArgs.sections,
-				NoTest:            buildArgs.noTest,
-				NoHTTPS:           noHTTPS,
-				LibraryURL:        buildArgs.libraryURL,
-				LibraryAuthToken:  authToken,
-				FakerootPath:      fakerootPath,
-				KeyServerOpts:     ko,
-				OCIAuthConfig:     authConf,
-				DockerDaemonHost:  dockerHost,
-				EncryptionKeyInfo: keyInfo,
-				FixPerms:          buildArgs.fixPerms,
-				SandboxTarget:     sandboxTarget,
-				DataPartition:     dataPartition,
-				MksquashfsArgs:    buildArgs.mksquashfsArgs,
-				Binds:             buildArgs.bindPaths,
-				Unprivilege:       unprivilege,
-				ReqAuthFile:       reqAuthFile,
-				Arch:              arch,
-				Platform:          *dp,
-			},
-		})
+	config := build.Config{
+		Dest:      dst,
+		Format:    buildFormat,
+		NoCleanUp: buildArgs.noCleanUp,
+		Opts: types.Options{
+			ImgCache:          imgCache,
+			TmpDir:            tmpDir,
+			NoCache:           disableCache,
+			Update:            buildArgs.update,
+			Force:             forceOverwrite,
+			Sections:          buildArgs.sections,
+			NoTest:            buildArgs.noTest,
+			NoHTTPS:           noHTTPS,
+			LibraryURL:        buildArgs.libraryURL,
+			LibraryAuthToken:  authToken,
+			FakerootPath:      fakerootPath,
+			KeyServerOpts:     ko,
+			OCIAuthConfig:     authConf,
+			DockerDaemonHost:  dockerHost,
+			EncryptionKeyInfo: keyInfo,
+			FixPerms:          buildArgs.fixPerms,
+			SandboxTarget:     sandboxTarget,
+			DataPartition:     dataPartition,
+			MksquashfsArgs:    buildArgs.mksquashfsArgs,
+			Binds:             buildArgs.bindPaths,
+			Unprivilege:       unprivilege,
+			ReqAuthFile:       reqAuthFile,
+			Arch:              arch,
+			Platform:          *dp,
+			Reproducible:      buildArgs.reproducible,
+		},
+	}
+	b, err := build.New(defs, config)
 	if err != nil {
 		sylog.Fatalf("Unable to create build: %v", err)
 	}